requirement.txt
absl-py==2.1.0
asgiref==3.9.1
asttokens==2.4.1
attrs==23.2.0
cffi==1.16.0
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
Django==5.2.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
executing==2.0.1
flatbuffers==24.3.25
fonttools==4.51.0
ipykernel==6.29.4
ipython==8.24.0
jax==0.4.28
jaxlib==0.4.28
jedi==0.19.1
joblib==1.4.2
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mediapipe==0.10.14
ml-dtypes==0.4.0
MouseInfo==0.1.3
nest-asyncio==1.6.0
numpy==1.26.4
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==24.0
pandas==2.2.3
parso==0.8.4
pillow==10.3.0
platformdirs==4.2.1
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
PyAutoGUI==0.9.54
pycparser==2.22
pycryptodome==3.23.0
PyGetWindow==0.0.9
Pygments==2.17.2
PyJWT==2.9.0
PyMsgBox==1.0.9
pyparsing==3.1.2
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.9.0.post0
pytweening==1.2.0
pytz==2024.2
pywin32==306
pyzmq==26.0.2
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
sounddevice==0.4.6
sqlparse==0.5.3
stack-data==0.6.3
threadpoolctl==3.5.0
tornado==6.4
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.2
wcwidth==0.2.13
dj-database-url
psycopg2-binary
